Improve styling to allow flexible overrides
This commit changes targeted styling of `a` element to a class-based
styling. This way, this is style is reusable on other elements and
another `<a>` element can be used inside a `post-item`.

Otherwise, if someone would need to use `<a>` element in a post item,
then he would get all other styles (such as `display:flex`) on every
`<a>` which is unexpected.

This commit ensures that the style is abstracted from the element type
itself, giving more flexibility when overriding the theme, and increases
maintainability by making the intent more clear.